cdk/requirements.txt
annotated-types==0.7.0
anthropic==0.28.1
anyio==4.4.0
attrs==23.2.0
aws-cdk-lib==2.146.0
aws-cdk.asset-awscli-v1==2.2.202
aws-cdk.asset-kubectl-v20==2.1.2
aws-cdk.asset-node-proxy-agent-v6==2.0.3
aws-cdk.aws-cognito-identitypool-alpha==2.146.0a0
boto3==1.34.129
botocore==1.34.129
cattrs==23.2.3
certifi==2024.6.2
charset-normalizer==3.3.2
constructs==10.3.0
distro==1.9.0
Events==0.5
filelock==3.15.3
fsspec==2024.6.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.4
idna==3.7
importlib_resources==6.4.0
jiter==0.4.2
jmespath==1.0.1
jsii==1.100.0
opensearch-py==2.6.0
packaging==24.1
publication==0.0.3
pydantic==2.7.4
pydantic_core==2.18.4
python-dateutil==2.9.0.post0
PyYAML==6.0.1
requests==2.32.3
s3transfer==0.10.1
six==1.16.0
sniffio==1.3.1
tokenizers==0.19.1
tqdm==4.66.4
typeguard==2.13.3
typing_extensions==4.12.2
urllib3==2.2.2
